perf: separate `SELECT` components from visual editor
